refactor: send `height` as `double` value (#129)
## 📜 Description

Return `height` value as `double` instead of `integer`.

## 💡 Motivation and Context

Rounded values can create an effect of "laggy" animations. It's hard to
notice when keyboard simply opens and closes, but it's easy to notice
when keyboard becomes interactive (i. e. position can be controlled via
gesture). To fix this effect we need to send as precise as possible
values.

## 📢 Changelog

### JS
- changed spec of `KeyboardViewController` (now both `progress` and
`double` are returned as `double`);

### iOS
- send `height` as double on Fabric architecture;

### Android
- send `height` as double;
- fixed Android studio warnings;

## 🤔 How Has This Been Tested?

Tested both (fabric & paper) on:
- Pixel 7 Pro (real device, Android 13);
- iPhone 14 Pro (emulator, iOS 16.2).

## 📸 Screenshots (if appropriate):

|Before|After|
|-------|-----|
|<video
src="https://user-images.githubusercontent.com/22820318/227630300-98ca4ac9-6c15-4d3c-b705-6d8cc87d1fe1.mp4"
/>|<video
src="https://user-images.githubusercontent.com/22820318/227630393-4923b68b-245c-41f6-a072-90a2d92a2ee1.mp4"
/>|

## 📝 Checklist

- [x] CI successfully passed

+++ android/src/main/java/com/reactnativekeyboardcontroller/events/KeyboardTransitionEvent.kt
@@ -4,7 +4,7 @@ import com.facebook.react.bridge.Arguments
 import com.facebook.react.uimanager.events.Event
 import com.facebook.react.uimanager.events.RCTEventEmitter
 
-class KeyboardTransitionEvent(viewId: Int, private val event: String, private val height: Int, private val progress: Double) : Event<KeyboardTransitionEvent>(viewId) {
+class KeyboardTransitionEvent(viewId: Int, private val event: String, private val height: Double, private val progress: Double) : Event<KeyboardTransitionEvent>(viewId) {
   override fun getEventName() = event
 
   // TODO: All events for a given view can be coalesced?
@@ -13,7 +13,7 @@ class KeyboardTransitionEvent(viewId: Int, private val event: String, private va
   override fun dispatch(rctEventEmitter: RCTEventEmitter) {
     val map = Arguments.createMap()
     map.putDouble("progress", progress)
-    map.putInt("height", height)
+    map.putDouble("height", height)
     rctEventEmitter.receiveEvent(viewTag, eventName, map)
   }
 }
